requirements.txt
streamlit
crewai
faiss-cpu
langchain
langchain-community
sentence-transformers
pypdf
numpy
python-dotenv